+++ NEWS.md
@@ -1,13 +1,15 @@
 # rAmCharts4 1.0.0.9000 (2020-10-xx)
 
+* Updated amcharts4.
 * New function `updateAmBarChart`, to update the data of a bar chart in Shiny 
 (vertical, horizontal, or radial barchart).
+* New chart: stacked bar chart.
 
 
 # rAmCharts4 1.0.0 (2020-10-01)
 
-* Updated amcharts4
-* New chart: gauge chart
+* Updated amcharts4.
+* New chart: gauge chart.
 
 
 # rAmCharts4 0.1.0 (2020-08-24)